requirements.txt
Flask==1.0.2
flask-caching==1.4.0
Flask-Cors==3.0.2
marshmallow==3.0.0b20
boto3==1.4.4
flasgger==0.9.0

Flask-Login==0.4.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
Flask-User==0.6.9

Flask-Marshmallow==0.9.0
Marshmallow-SQLAlchemy==0.15.0

marshmallow-enum==1.4.1